requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pillow==10.3.0
torch==2.3.0
torchvision==0.18.0
git+https://github.com/openai/CLIP.git
python-dotenv==1.0.1
pydantic[email]==2.7.1
pydantic-settings==2.3.0
